+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Fixed
 
 - `sandstone/VirtualList` jumping focus when keys pressed in a row
+- `sandstone/FixedPopupPanels` to support accessibility properly
 - `sandstone/TabLayout` performance when focusing items in the layout
 
 ## [1.0.0-beta.6] - 2020-06-08